feat(web): add hooks to create issue
