+++ apps/web/src/hooks/data/use-projects.tsx
@@ -1,0 +1,9 @@
+import { api } from '@/lib/trpc/react'
+
+/**
+ * The hook to get all projects.
+ * @returns The projects.
+ */
+export function useProjects() {
+  return api.project.get_projects.useQuery()
+}
